requirements.txt
# API dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3

# Vector search dependencies
numpy==1.24.3
scipy==1.11.3
matplotlib==3.8.0
tqdm==4.66.1

# Configuration
python-dotenv==1.0.0

# Utilities
requests==2.31.0
